requirements.txt
# Core
streamlit==1.32.0
python-dotenv==1.0.0
requests==2.31.0
httpx==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# YandexGPT (Yandex Cloud)
yandexcloud==0.248.0
grpcio==1.60.0
grpcio-tools==1.60.0

# GigaChat (Sber)
gigachat==0.1.28

# Document Parsing
pypdf==4.0.1
pdfplumber==0.10.4
python-docx==1.1.0
docxtpl==0.16.7
pytesseract==0.3.10
Pillow==10.2.0

# Text Processing
beautifulsoup4==4.12.3
lxml==5.1.0
rank-bm25==0.2.2
sentence-transformers==2.3.1

# Memory & Cache
redis==5.0.1
diskcache==5.6.3

# Utilities
python-magic-bin==0.4.14
chardet==5.2.0
tenacity==8.2.3

# Reporting
reportlab==4.0.8
markdown==3.5.2

# Logging
loguru==0.7.2
